// ui/src/app/models/customer.model.ts
export interface Customer {
    id?: number;
    netsuite_customer_id?: string;
    company_name?: string;
    web_address?: string;
    email_address?: string;
    phone?: string;
    alt_phone?: string;
    fax_number?: string;
    address?: string;
    can_login?: boolean;
    password?: string;
    last_activity?: Date;
     
    
}